Continuing to refactor tests
