+++ CMakeLists.txt
@@ -340,6 +340,7 @@ ELSE(DL_LIBRARY)
 ENDIF(DL_LIBRARY)
 
 IF(BUILD_TESTING)
+    INCLUDE_DIRECTORIES(${CMAKE_SOURCE_DIR}/tests)
     ADD_SUBDIRECTORY(platforms/reference/tests)
 ENDIF(BUILD_TESTING)
