+++ platforms/opencl/tests/TestOpenCLCustomBondForce.cpp
@@ -6,7 +6,7 @@
  * Biological Structures at Stanford, funded under the NIH Roadmap for        *
  * Medical Research, grant U54 GM072970. See https://simtk.org.               *
  *                                                                            *
- * Portions copyright (c) 2008-2009 Stanford University and the Authors.      *
+ * Portions copyright (c) 2008-2015 Stanford University and the Authors.      *
  * Authors: Peter Eastman                                                     *
  * Contributors:                                                              *
  *                                                                            *
@@ -29,111 +29,8 @@
  * USE OR OTHER DEALINGS IN THE SOFTWARE.                                     *
  * -------------------------------------------------------------------------- */
 
-/**
- * This tests the OpenCL implementation of CustomBondForce.
- */
-
-#include "openmm/internal/AssertionUtilities.h"
-#include "openmm/Context.h"
-#include "OpenCLPlatform.h"
-#include "openmm/CustomBondForce.h"
-#include "openmm/System.h"
-#include "openmm/VerletIntegrator.h"
-#include "SimTKOpenMMRealType.h"
-#include <iostream>
-#include <vector>
-
-using namespace OpenMM;
-using namespace std;
-
-static OpenCLPlatform platform;
-
-const double TOL = 1e-5;
-
-void testBonds() {
-    System system;
-    system.addParticle(1.0);
-    system.addParticle(1.0);
-    system.addParticle(1.0);
-    VerletIntegrator integrator(0.01);
-    CustomBondForce* forceField = new CustomBondForce("scale*k*(r-r0)^2");
-    forceField->addPerBondParameter("r0");
-    forceField->addPerBondParameter("k");
-    forceField->addGlobalParameter("scale", 0.5);
-    vector<double> parameters(2);
-    parameters[0] = 1.5;
-    parameters[1] = 0.8;
-    forceField->addBond(0, 1, parameters);
-    parameters[0] = 1.2;
-    parameters[1] = 0.7;
-    forceField->addBond(1, 2, parameters);
-    system.addForce(forceField);
-    Context context(system, integrator, platform);
-    vector<Vec3> positions(3);
-    positions[0] = Vec3(0, 2, 0);
-    positions[1] = Vec3(0, 0, 0);
-    positions[2] = Vec3(1, 0, 0);
-    context.setPositions(positions);
-    State state = context.getState(State::Forces | State::Energy);
-    {
-        const vector<Vec3>& forces = state.getForces();
-        ASSERT_EQUAL_VEC(Vec3(0, -0.8*0.5, 0), forces[0], TOL);
-        ASSERT_EQUAL_VEC(Vec3(0.7*0.2, 0, 0), forces[2], TOL);
-        ASSERT_EQUAL_VEC(Vec3(-forces[0][0]-forces[2][0], -forces[0][1]-forces[2][1], -forces[0][2]-forces[2][2]), forces[1], TOL);
-        ASSERT_EQUAL_TOL(0.5*0.8*0.5*0.5 + 0.5*0.7*0.2*0.2, state.getPotentialEnergy(), TOL);
-    }
-    
-    // Try changing the bond parameters and make sure it's still correct.
-    
-    parameters[0] = 1.6;
-    parameters[1] = 0.9;
-    forceField->setBondParameters(0, 0, 1, parameters);
-    parameters[0] = 1.3;
-    parameters[1] = 0.8;
-    forceField->setBondParameters(1, 1, 2, parameters);
-    forceField->updateParametersInContext(context);
-    state = context.getState(State::Forces | State::Energy);
-    {
-        const vector<Vec3>& forces = state.getForces();
-        ASSERT_EQUAL_VEC(Vec3(0, -0.9*0.4, 0), forces[0], TOL);
-        ASSERT_EQUAL_VEC(Vec3(0.8*0.3, 0, 0), forces[2], TOL);
-        ASSERT_EQUAL_VEC(Vec3(-forces[0][0]-forces[2][0], -forces[0][1]-forces[2][1], -forces[0][2]-forces[2][2]), forces[1], TOL);
-        ASSERT_EQUAL_TOL(0.5*0.9*0.4*0.4 + 0.5*0.8*0.3*0.3, state.getPotentialEnergy(), TOL);
-    }
-}
-
-void testManyParameters() {
-    System system;
-    system.addParticle(1.0);
-    system.addParticle(1.0);
-    VerletIntegrator integrator(0.01);
-    CustomBondForce* forceField = new CustomBondForce("(a+b+c+d+e+f+g+h+i)*r");
-    forceField->addPerBondParameter("a");
-    forceField->addPerBondParameter("b");
-    forceField->addPerBondParameter("c");
-    forceField->addPerBondParameter("d");
-    forceField->addPerBondParameter("e");
-    forceField->addPerBondParameter("f");
-    forceField->addPerBondParameter("g");
-    forceField->addPerBondParameter("h");
-    forceField->addPerBondParameter("i");
-    vector<double> parameters(forceField->getNumPerBondParameters());
-    for (int i = 0; i < (int) parameters.size(); i++)
-        parameters[i] = i;
-    forceField->addBond(0, 1, parameters);
-    system.addForce(forceField);
-    Context context(system, integrator, platform);
-    vector<Vec3> positions(2);
-    positions[0] = Vec3(0, 0, 0);
-    positions[1] = Vec3(0, 2.5, 0);
-    context.setPositions(positions);
-    State state = context.getState(State::Forces | State::Energy);
-    const vector<Vec3>& forces = state.getForces();
-    double f = 1+2+3+4+5+6+7+8;
-    ASSERT_EQUAL_VEC(Vec3(0, f, 0), forces[0], TOL);
-    ASSERT_EQUAL_VEC(Vec3(0, -f, 0), forces[1], TOL);
-    ASSERT_EQUAL_TOL(f*2.5, state.getPotentialEnergy(), TOL);
-}
+#include "OpenCLTests.h"
+#include "TestCustomBondForce.h"
 
 void testParallelComputation() {
     System system;
@@ -164,19 +61,6 @@ void testParallelComputation() {
         ASSERT_EQUAL_VEC(state1.getForces()[i], state2.getForces()[i], 1e-5);
 }
 
-int main(int argc, char* argv[]) {
-    try {
-        if (argc > 1)
-            platform.setPropertyDefaultValue("OpenCLPrecision", string(argv[1]));
-        testBonds();
-        testManyParameters();
-        testParallelComputation();
-    }
-    catch(const exception& e) {
-        cout << "exception: " << e.what() << endl;
-        return 1;
-    }
-    cout << "Done" << endl;
-    return 0;
+void runPlatformTests() {
+    testParallelComputation();
 }
-
