+++ platforms/opencl/tests/TestOpenCLRBTorsionForce.cpp
@@ -29,88 +29,8 @@
  * USE OR OTHER DEALINGS IN THE SOFTWARE.                                     *
  * -------------------------------------------------------------------------- */
 
-/**
- * This tests the OpenCL implementation of RBTorsionForce.
- */
-
-#include "openmm/internal/AssertionUtilities.h"
-#include "openmm/Context.h"
-#include "OpenCLPlatform.h"
-#include "openmm/RBTorsionForce.h"
-#include "openmm/System.h"
-#include "openmm/VerletIntegrator.h"
-#include "SimTKOpenMMRealType.h"
-#include <iostream>
-#include <vector>
-
-using namespace OpenMM;
-using namespace std;
-
-static OpenCLPlatform platform;
-
-const double TOL = 1e-5;
-
-void testRBTorsions() {
-    System system;
-    system.addParticle(1.0);
-    system.addParticle(1.0);
-    system.addParticle(1.0);
-    system.addParticle(1.0);
-    VerletIntegrator integrator(0.01);
-    RBTorsionForce* forceField = new RBTorsionForce();
-    forceField->addTorsion(0, 1, 2, 3, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6);
-    system.addForce(forceField);
-    Context context(system, integrator, platform);
-    vector<Vec3> positions(4);
-    positions[0] = Vec3(0, 1, 0);
-    positions[1] = Vec3(0, 0, 0);
-    positions[2] = Vec3(1, 0, 0);
-    positions[3] = Vec3(1, 1, 1);
-    context.setPositions(positions);
-    State state = context.getState(State::Forces | State::Energy);
-    {
-        const vector<Vec3>& forces = state.getForces();
-        double psi = 0.25*PI_M - PI_M;
-        double torque = 0.0;
-        for (int i = 1; i < 6; ++i) {
-            double c = 0.1*(i+1);
-            torque += -c*i*std::pow(std::cos(psi), i-1)*std::sin(psi);
-        }
-        ASSERT_EQUAL_VEC(Vec3(0, 0, torque), forces[0], TOL);
-        ASSERT_EQUAL_VEC(Vec3(0, 0.5*torque, -0.5*torque), forces[3], TOL);
-        ASSERT_EQUAL_VEC(Vec3(forces[0][0]+forces[1][0]+forces[2][0]+forces[3][0], forces[0][1]+forces[1][1]+forces[2][1]+forces[3][1], forces[0][2]+forces[1][2]+forces[2][2]+forces[3][2]), Vec3(0, 0, 0), TOL);
-        double energy = 0.0;
-        for (int i = 0; i < 6; ++i) {
-            double c = 0.1*(i+1);
-            energy += c*std::pow(std::cos(psi), i);
-        }
-        ASSERT_EQUAL_TOL(energy, state.getPotentialEnergy(), TOL);
-    }
-    
-    // Try changing the torsion parameters and make sure it's still correct.
-    
-    forceField->setTorsionParameters(0, 0, 1, 2, 3, 0.11, 0.22, 0.33, 0.44, 0.55, 0.66);
-    forceField->updateParametersInContext(context);
-    state = context.getState(State::Forces | State::Energy);
-    {
-        const vector<Vec3>& forces = state.getForces();
-        double psi = 0.25*PI_M - PI_M;
-        double torque = 0.0;
-        for (int i = 1; i < 6; ++i) {
-            double c = 0.11*(i+1);
-            torque += -c*i*std::pow(std::cos(psi), i-1)*std::sin(psi);
-        }
-        ASSERT_EQUAL_VEC(Vec3(0, 0, torque), forces[0], TOL);
-        ASSERT_EQUAL_VEC(Vec3(0, 0.5*torque, -0.5*torque), forces[3], TOL);
-        ASSERT_EQUAL_VEC(Vec3(forces[0][0]+forces[1][0]+forces[2][0]+forces[3][0], forces[0][1]+forces[1][1]+forces[2][1]+forces[3][1], forces[0][2]+forces[1][2]+forces[2][2]+forces[3][2]), Vec3(0, 0, 0), TOL);
-        double energy = 0.0;
-        for (int i = 0; i < 6; ++i) {
-            double c = 0.11*(i+1);
-            energy += c*std::pow(std::cos(psi), i);
-        }
-        ASSERT_EQUAL_TOL(energy, state.getPotentialEnergy(), TOL);
-    }
-}
+#include "OpenCLTests.h"
+#include "TestRBTorsionForce.h"
 
 void testParallelComputation() {
     System system;
@@ -140,18 +60,6 @@ void testParallelComputation() {
         ASSERT_EQUAL_VEC(state1.getForces()[i], state2.getForces()[i], 1e-5);
 }
 
-int main(int argc, char* argv[]) {
-    try {
-        if (argc > 1)
-            platform.setPropertyDefaultValue("OpenCLPrecision", string(argv[1]));
-        testRBTorsions();
-        testParallelComputation();
-    }
-    catch(const exception& e) {
-        cout << "exception: " << e.what() << endl;
-        return 1;
-    }
-    cout << "Done" << endl;
-    return 0;
+void runPlatformTests() {
+    testParallelComputation();
 }
-
