+++ platforms/reference/tests/TestReferenceCustomAngleForce.cpp
@@ -6,7 +6,7 @@
  * Biological Structures at Stanford, funded under the NIH Roadmap for        *
  * Medical Research, grant U54 GM072970. See https://simtk.org.               *
  *                                                                            *
- * Portions copyright (c) 2008-2010 Stanford University and the Authors.      *
+ * Portions copyright (c) 2015 Stanford University and the Authors.           *
  * Authors: Peter Eastman                                                     *
  * Contributors:                                                              *
  *                                                                            *
@@ -29,122 +29,9 @@
  * USE OR OTHER DEALINGS IN THE SOFTWARE.                                     *
  * -------------------------------------------------------------------------- */
 
-/**
- * This tests the reference implementation of CustomAngleForce.
- */
 
-#include "openmm/internal/AssertionUtilities.h"
-#include "openmm/Context.h"
-#include "ReferencePlatform.h"
-#include "openmm/CustomAngleForce.h"
-#include "openmm/HarmonicAngleForce.h"
-#include "openmm/System.h"
-#include "openmm/VerletIntegrator.h"
-#include "sfmt/SFMT.h"
-#include <iostream>
-#include <vector>
+#include "ReferenceTests.h"
+#include "TestCustomAngleForce.h"
 
-using namespace OpenMM;
-using namespace std;
-
-ReferencePlatform platform;
-
-const double TOL = 1e-5;
-
-void testAngles() {
-    // Create a system using a CustomAngleForce.
-
-    System customSystem;
-    customSystem.addParticle(1.0);
-    customSystem.addParticle(1.0);
-    customSystem.addParticle(1.0);
-    customSystem.addParticle(1.0);
-    CustomAngleForce* custom = new CustomAngleForce("scale*k*(theta-theta0)^2");
-    custom->addPerAngleParameter("theta0");
-    custom->addPerAngleParameter("k");
-    custom->addGlobalParameter("scale", 0.5);
-    vector<double> parameters(2);
-    parameters[0] = 1.5;
-    parameters[1] = 0.8;
-    custom->addAngle(0, 1, 2, parameters);
-    parameters[0] = 2.0;
-    parameters[1] = 0.5;
-    custom->addAngle(1, 2, 3, parameters);
-    customSystem.addForce(custom);
-    ASSERT(!custom->usesPeriodicBoundaryConditions());
-    ASSERT(!customSystem.usesPeriodicBoundaryConditions());
-
-    // Create an identical system using a HarmonicAngleForce.
-
-    System harmonicSystem;
-    harmonicSystem.addParticle(1.0);
-    harmonicSystem.addParticle(1.0);
-    harmonicSystem.addParticle(1.0);
-    harmonicSystem.addParticle(1.0);
-    HarmonicAngleForce* harmonic = new HarmonicAngleForce();
-    harmonic->addAngle(0, 1, 2, 1.5, 0.8);
-    harmonic->addAngle(1, 2, 3, 2.0, 0.5);
-    harmonicSystem.addForce(harmonic);
-
-    // Set the atoms in various positions, and verify that both systems give identical forces and energy.
-
-    OpenMM_SFMT::SFMT sfmt;
-    init_gen_rand(0, sfmt);
-
-    vector<Vec3> positions(4);
-    VerletIntegrator integrator1(0.01);
-    VerletIntegrator integrator2(0.01);
-    Context c1(customSystem, integrator1, platform);
-    Context c2(harmonicSystem, integrator2, platform);
-    for (int i = 0; i < 10; i++) {
-        for (int j = 0; j < (int) positions.size(); j++)
-            positions[j] = Vec3(5.0*genrand_real2(sfmt), 5.0*genrand_real2(sfmt), 5.0*genrand_real2(sfmt));
-        c1.setPositions(positions);
-        c2.setPositions(positions);
-        State s1 = c1.getState(State::Forces | State::Energy);
-        State s2 = c2.getState(State::Forces | State::Energy);
-        const vector<Vec3>& forces = s1.getForces();
-        for (int i = 0; i < customSystem.getNumParticles(); i++)
-            ASSERT_EQUAL_VEC(s1.getForces()[i], s2.getForces()[i], TOL);
-        ASSERT_EQUAL_TOL(s1.getPotentialEnergy(), s2.getPotentialEnergy(), TOL);
-    }
-    
-    // Try changing the angle parameters and make sure it's still correct.
-    
-    parameters[0] = 1.6;
-    parameters[1] = 0.9;
-    custom->setAngleParameters(0, 0, 1, 2, parameters);
-    parameters[0] = 2.1;
-    parameters[1] = 0.6;
-    custom->setAngleParameters(1, 1, 2, 3, parameters);
-    custom->updateParametersInContext(c1);
-    harmonic->setAngleParameters(0, 0, 1, 2, 1.6, 0.9);
-    harmonic->setAngleParameters(1, 1, 2, 3, 2.1, 0.6);
-    harmonic->updateParametersInContext(c2);
-    {
-        for (int j = 0; j < (int) positions.size(); j++)
-            positions[j] = Vec3(5.0*genrand_real2(sfmt), 5.0*genrand_real2(sfmt), 5.0*genrand_real2(sfmt));
-        c1.setPositions(positions);
-        c2.setPositions(positions);
-        State s1 = c1.getState(State::Forces | State::Energy);
-        State s2 = c2.getState(State::Forces | State::Energy);
-        const vector<Vec3>& forces = s1.getForces();
-        for (int i = 0; i < customSystem.getNumParticles(); i++)
-            ASSERT_EQUAL_VEC(s1.getForces()[i], s2.getForces()[i], TOL);
-        ASSERT_EQUAL_TOL(s1.getPotentialEnergy(), s2.getPotentialEnergy(), TOL);
-    }
+void runPlatformTests() {
 }
-
-int main() {
-    try {
-        testAngles();
-    }
-    catch(const exception& e) {
-        cout << "exception: " << e.what() << endl;
-        return 1;
-    }
-    cout << "Done" << endl;
-    return 0;
-}
-
-
