+++ tests/TestRBTorsionForce.h
@@ -6,7 +6,7 @@
  * Biological Structures at Stanford, funded under the NIH Roadmap for        *
  * Medical Research, grant U54 GM072970. See https://simtk.org.               *
  *                                                                            *
- * Portions copyright (c) 2008-2015 Stanford University and the Authors.s      *
+ * Portions copyright (c) 2008-2015 Stanford University and the Authors.      *
  * Authors: Peter Eastman                                                     *
  * Contributors:                                                              *
  *                                                                            *
@@ -31,7 +31,7 @@
 
 #include "openmm/internal/AssertionUtilities.h"
 #include "openmm/Context.h"
-#include "openmm/HarmonicAngleForce.h"
+#include "openmm/RBTorsionForce.h"
 #include "openmm/System.h"
 #include "openmm/VerletIntegrator.h"
 #include "SimTKOpenMMRealType.h"
@@ -43,16 +43,15 @@ using namespace std;
 
 const double TOL = 1e-5;
 
-void testAngles() {
+void testRBTorsions() {
     System system;
     system.addParticle(1.0);
     system.addParticle(1.0);
     system.addParticle(1.0);
     system.addParticle(1.0);
     VerletIntegrator integrator(0.01);
-    HarmonicAngleForce* forceField = new HarmonicAngleForce();
-    forceField->addAngle(0, 1, 2, PI_M/3, 1.1);
-    forceField->addAngle(1, 2, 3, PI_M/2, 1.2);
+    RBTorsionForce* forceField = new RBTorsionForce();
+    forceField->addTorsion(0, 1, 2, 3, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6);
     system.addForce(forceField);
     ASSERT(!forceField->usesPeriodicBoundaryConditions());
     ASSERT(!system.usesPeriodicBoundaryConditions());
@@ -61,35 +60,50 @@ void testAngles() {
     positions[0] = Vec3(0, 1, 0);
     positions[1] = Vec3(0, 0, 0);
     positions[2] = Vec3(1, 0, 0);
-    positions[3] = Vec3(2, 1, 0);
+    positions[3] = Vec3(1, 1, 1);
     context.setPositions(positions);
     State state = context.getState(State::Forces | State::Energy);
     {
         const vector<Vec3>& forces = state.getForces();
-        double torque1 = 1.1*PI_M/6;
-        double torque2 = 1.2*PI_M/4;
-        ASSERT_EQUAL_VEC(Vec3(torque1, 0, 0), forces[0], TOL);
-        ASSERT_EQUAL_VEC(Vec3(-0.5*torque2, 0.5*torque2, 0), forces[3], TOL); // reduced by sqrt(2) due to the bond length, another sqrt(2) due to the angle
+        double psi = 0.25*PI_M - PI_M;
+        double torque = 0.0;
+        for (int i = 1; i < 6; ++i) {
+            double c = 0.1*(i+1);
+            torque += -c*i*std::pow(std::cos(psi), i-1)*std::sin(psi);
+        }
+        ASSERT_EQUAL_VEC(Vec3(0, 0, torque), forces[0], TOL);
+        ASSERT_EQUAL_VEC(Vec3(0, 0.5*torque, -0.5*torque), forces[3], TOL);
         ASSERT_EQUAL_VEC(Vec3(forces[0][0]+forces[1][0]+forces[2][0]+forces[3][0], forces[0][1]+forces[1][1]+forces[2][1]+forces[3][1], forces[0][2]+forces[1][2]+forces[2][2]+forces[3][2]), Vec3(0, 0, 0), TOL);
-        ASSERT_EQUAL_TOL(0.5*1.1*(PI_M/6)*(PI_M/6) + 0.5*1.2*(PI_M/4)*(PI_M/4), state.getPotentialEnergy(), TOL);
+        double energy = 0.0;
+        for (int i = 0; i < 6; ++i) {
+            double c = 0.1*(i+1);
+            energy += c*std::pow(std::cos(psi), i);
+        }
+        ASSERT_EQUAL_TOL(energy, state.getPotentialEnergy(), TOL);
     }
     
-    // Try changing the angle parameters and make sure it's still correct.
+    // Try changing the torsion parameters and make sure it's still correct.
     
-    forceField->setAngleParameters(0, 0, 1, 2, PI_M/3.1, 1.3);
-    forceField->setAngleParameters(1, 1, 2, 3, PI_M/2.1, 1.4);
+    forceField->setTorsionParameters(0, 0, 1, 2, 3, 0.11, 0.22, 0.33, 0.44, 0.55, 0.66);
     forceField->updateParametersInContext(context);
     state = context.getState(State::Forces | State::Energy);
     {
         const vector<Vec3>& forces = state.getForces();
-        double dtheta1 = (PI_M/2)-(PI_M/3.1);
-        double dtheta2 = (3*PI_M/4)-(PI_M/2.1);
-        double torque1 = 1.3*dtheta1;
-        double torque2 = 1.4*dtheta2;
-        ASSERT_EQUAL_VEC(Vec3(torque1, 0, 0), forces[0], TOL);
-        ASSERT_EQUAL_VEC(Vec3(-0.5*torque2, 0.5*torque2, 0), forces[3], TOL); // reduced by sqrt(2) due to the bond length, another sqrt(2) due to the angle
+        double psi = 0.25*PI_M - PI_M;
+        double torque = 0.0;
+        for (int i = 1; i < 6; ++i) {
+            double c = 0.11*(i+1);
+            torque += -c*i*std::pow(std::cos(psi), i-1)*std::sin(psi);
+        }
+        ASSERT_EQUAL_VEC(Vec3(0, 0, torque), forces[0], TOL);
+        ASSERT_EQUAL_VEC(Vec3(0, 0.5*torque, -0.5*torque), forces[3], TOL);
         ASSERT_EQUAL_VEC(Vec3(forces[0][0]+forces[1][0]+forces[2][0]+forces[3][0], forces[0][1]+forces[1][1]+forces[2][1]+forces[3][1], forces[0][2]+forces[1][2]+forces[2][2]+forces[3][2]), Vec3(0, 0, 0), TOL);
-        ASSERT_EQUAL_TOL(0.5*1.3*dtheta1*dtheta1 + 0.5*1.4*dtheta2*dtheta2, state.getPotentialEnergy(), TOL);
+        double energy = 0.0;
+        for (int i = 0; i < 6; ++i) {
+            double c = 0.11*(i+1);
+            energy += c*std::pow(std::cos(psi), i);
+        }
+        ASSERT_EQUAL_TOL(energy, state.getPotentialEnergy(), TOL);
     }
 }
 
@@ -98,7 +112,7 @@ void runPlatformTests();
 int main(int argc, char* argv[]) {
     try {
         initializeTests(argc, argv);
-        testAngles();
+        testRBTorsions();
         runPlatformTests();
     }
     catch(const exception& e) {
